services/alert/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
kafka-python==2.0.2
cryptography==41.0.7